chore: normalize demo styles
