+++ demo/src/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import ReactDOM from 'react-dom'
-import './styles.css'
+import './index.css'
 import App from './App'
 
 ReactDOM.render(<App />, document.getElementById('root'))
