+++ demo/src/sandboxes/card/src/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
+import './index.css'
 
 const rootElement = document.getElementById('root')
 ReactDOM.render(
